+++ packages/core/rpc_dart/test/transports/receive_path_hardening_test.dart
@@ -257,5 +257,4 @@ void main() {
       expect(frames.single.methodPath, equals('/svc/Method'));
     });
   });
-
 }